Did some thiiiiings.
Made the transaction edit a modal instead. Well modal-like. Still a component, but motherfucking CSS'd the shit into a modal!!!

It's unnecessary currently having a modal component and then the editform inside it. Adds for, well it adds. adding is bad in coding. Rather subtract. Cause editForm is the only one using the modal component container.

The form is set-up to edit. Some things, problems problems problems:
- Date doesn't display.
- Switched form reactive to template forms. somewhere some data, something no works. Doesn't update from the initialization values (which can be viewed in the console by clicking save

That's enough for now.
Ciao