Timestamps Fixes #10
Added empty space to display should the running total not show in order to keep the buttons' position fixed

Fixed this time thing. The first section.

Not finding any errors with time displaying wrong. I changed to input for the angular DatePipe in the HTML for the ledgerView from the date variable to be date.toString(). No problems on PC.

Also some cleaning as always.

And did do something cool. But what? FInally removed the JSONPipe angular compiler has been bitching about.

A yeah, double saved every time a transaction was edited becuase of actions on (ngSubmit) and on (click) save button. Now only a single save